flatten packages in api, refactor in HttpExceptionInfo
