+++ codebase/api/api-common/src/main/java/com/jarqprog/commonapi/components/HttpExceptionDTO.java
@@ -1,0 +1,13 @@
+package com.jarqprog.commonapi.components;
+
+import com.jarqprog.commondomain.DTO;
+
+import java.util.Date;
+
+public interface HttpExceptionDTO extends DTO {
+
+    String uuid();
+    long httpStatus();
+    Date date();
+    String message();
+}
