--- codebase/api/api-common/src/main/java/com/jarqprog/commonapi/exceptions/exceptionInfo/ExceptionInfo.java
@@ -1,11 +1,0 @@
-package com.jarqprog.commonapi.exceptions.exceptionInfo;
-
-import java.time.LocalDateTime;
-
-public interface ExceptionInfo {
-
-    String getUuid();
-    long getHttpStatus();
-    LocalDateTime getDateTime();
-    String getMessage();
-}
